Plot species' activity patterns

This function visualizes species' activity patterns based on time-of-day data. It uses kernel density estimation to estimate activity density.

Usage

mm_plot_density(
  time,
  xscale = 24,
  xcenter = c("noon", "midnight"),
  n_grid = 128,
  kmax = 3,
  adjust = 1,
  rug = FALSE,
  linetype = 2,
  linecol = "gray10",
  linewidth = 1,
  rug_lentgh = 0.018,
  rug_color = "gray30",
  extend = "lightgrey",
  extend_alpha = 0.8,
  ...
)

Arguments

time

A numeric vector of time-of-day observations (in radians, 0 to \(2\pi\)).

xscale

A numeric value to scale the x-axis. Default is 24 for representing time in hours.

xcenter

A string indicating the center of the x-axis. Options are "noon" (default) or "midnight".

n_grid

An integer specifying the number of grid points for density estimation. Default is 128.

kmax

An integer indicating the maximum number of modes allowed in the activity pattern. Default is 3.

adjust

A numeric value to adjust the bandwidth of the kernel density estimation. Default is 1.

rug

A logical value indicating whether to include a rug plot of the observations. Default is FALSE.

linetype

A numeric specifying the line types. Default is c(1, 2).

linecol

A string specifying the colors of the density lines for species A and B. Default is c("gray10", "gray0").

linewidth

A numeric value specifying the line widths for species A and B density lines. Default is c(1, 1).

rug_lentgh

A numeric value specifying the length of the rug ticks. Default is 0.018 (in normalized plot coordinates).

rug_color

A string specifying the color of the rug ticks. Default is "gray30".

extend

A string specifying the color of the extended area beyond the activity period. Default is "lightgrey".

extend_alpha

A numeric value (0 to 1) for the transparency of the extended area. Default is 0.8.

...

Additional arguments passed to the geom_rug function.

Value

A ggplot object representing the activity density curves of the species.

Examples

 # Generate random data for two species
 set.seed(42)
 A <- runif(100, 0, 2 * pi)

 # Plot overlap with default settings
 mm_plot_density(A)
#> Error in mm_plot_density(A): object 'A' not found
 # Customize plot with specific colors and line types
 mm_plot_density(A, linecol = "gray10", linewidth = 0.8,
                 xcenter = "midnight", rug = TRUE,
                 rug_color = 'red', extend_alpha = 0)
#> Error in mm_plot_density(A, linecol = "gray10", linewidth = 0.8, xcenter = "midnight",     rug = TRUE, rug_color = "red", extend_alpha = 0): object 'A' not found