Transform Time to Solar Time Anchored to Sunrise and Sunset

This function converts local time to solar time based on the sunrise and sunset times for a given location. Solar time is a timekeeping system where the day is defined by the position of the sun in the sky, with sunrise marking the start of the day and sunset marking the end.

Usage

mm_solartime(
  data = NULL,
  date,
  longitude,
  latitude,
  crs = NULL,
  time_zone,
  ...,
  try_formats = c("%Y-%m-%d %H:%M:%OS", "%Y/%m/%d %H:%M:%OS",
    "%Y:%m:%d %H:%M:%OS", "%Y-%m-%d %H:%M", "%Y/%m/%d %H:%M",
    "%Y:%m:%d %H:%M", "%Y-%m-%d", "%Y/%m/%d", "%Y:%m:%d")
)

Arguments

data

A data frame containing the date, longitude, and latitude columns. If NULL, the function will use the date, longitude, and latitude parameters directly. Default is NULL.

date

A vector of date-time values or a column name in data representing the date-time values to be converted to solar time. This can be a character vector or a POSIXlt object.

longitude

A numeric vector or a column name in data representing the longitude of the location(s). Longitude should be in decimal degrees.

latitude

A numeric vector or a column name in data representing the latitude of the location(s). Latitude should be in decimal degrees.

crs

A coordinate reference system (CRS) string or object specifying the current CRS of the input coordinates. If provided, the function will transform the coordinates to longitude and latitude (WGS84). This is useful when the input coordinates are in a projected system (e.g., UTM). Default is NULL.

time_zone

A numeric vector representing the time zone offset(s) from UTC (in hours). If data is provided, this should match the number of unique locations in the data.

...

Additional arguments passed to as.POSIXlt for date parsing.

try_formats

A vector of date-time formats to try when parsing the date parameter. Default includes common date-time formats.

Value

A tibble with the following columns:

input: The original date-time values.

clock: The local clock time.

solar: The calculated solar time.

If data is provided, the tibble will also include the longitude and latitude columns.

Details

The function calculates solar time by first determining the sunrise and sunset times for the given location(s) and date(s). It then uses these times to anchor the solar time calculation. The solar time is computed by transforming the local clock time based on the position of the sun in the sky.

If data is provided, the function will process each unique location in the data and return a tibble with the solar time for each date-time value. If data is NULL, the function will process the date, longitude, and latitude parameters directly.

References

Rowcliffe, M. (2023). activity: Animal Activity Statistics. R package version 1.3.4. https://CRAN.R-project.org/package=activity

Examples

# Example with direct input
date <- c("2023-10-01 12:00:00", "2023-10-02 12:00:00")
longitude <- -122.4194
latitude <- 37.7749
time_zone <- -8
result <- mm_solartime(date = date, longitude = longitude, latitude = latitude,
time_zone = time_zone)

# Example with data frame
data <- data.frame(
  date_time = c("2023-10-01 12:00:00", "2023-10-02 12:00:00"),
  lon = c(-122.4194, -122.4194),
  lat = c(37.7749, 37.7749)
)
result <- mm_solartime(data = data, date = date_time, longitude = lon,
latitude = lat, time_zone = -8)
result
#> # A tibble: 2 × 5
#>     lon   lat date_time           clock solar
#>   <dbl> <dbl> <dttm>              <dbl> <dbl>
#> 1 -122.  37.8 2023-10-01 12:00:00  3.14  3.14
#> 2 -122.  37.8 2023-10-02 12:00:00  3.14  3.14